+++ config.cpp
@@ -22,7 +22,7 @@ class CfgMods
 		dir = "SentinelDM";
 		name = "Sentinel Deathmatch";
 		author = "Sentinel";
-		version = "0.1.6";
+		version = "0.1.7";
 		type = "mod";
 		dependencies[] = { "Game", "World", "Mission" };
 		class defs